// src/generated/models/Dga_module_configurationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dga_module_configurationsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dga_module_configurationsstatecode = keyof typeof Dga_module_configurationsstatecode;
export const Dga_module_configurationsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dga_module_configurationsstatuscode = keyof typeof Dga_module_configurationsstatuscode;

export interface Dga_module_configurationsBase {
  "dga_account@odata.bind"?: string;
  "dga_approver_team@odata.bind"?: string;
  dga_module_configurationid: string;
  "dga_module_type@odata.bind"?: string;
  dga_name: string;
  "dga_respondent_team@odata.bind"?: string;
  "dga_reviewer_team@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  "OwningBusinessUnit@odata.bind"?: string;
  statecode: Dga_module_configurationsstatecode;
  statuscode?: Dga_module_configurationsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dga_module_configurations extends Dga_module_configurationsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dga_accountname?: string;
  dga_accountyominame?: string;
  dga_approver_teamname?: string;
  dga_approver_teamyominame?: string;
  dga_module_typename?: string;
  dga_respondent_teamname?: string;
  dga_respondent_teamyominame?: string;
  dga_reviewer_teamname?: string;
  dga_reviewer_teamyominame?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  dga_account?: object;
  _dga_account_value?: string;
  dga_approver_team?: object;
  _dga_approver_team_value?: string;
  dga_module_type?: object;
  _dga_module_type_value?: string;
  dga_respondent_team?: object;
  _dga_respondent_team_value?: string;
  dga_reviewer_team?: object;
  _dga_reviewer_team_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}